apps/resume_evaluation_service/requirements.txt
openai==1.99.5
pydantic==2.11.7
fastapi==0.116.1
langchain==0.3.27
langchain-openai==0.3.29
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-google-genai==2.1.9
langgraph==0.6.4
pypdf==5.9.0
docx2txt==0.9
boto3==1.40.1
python-dotenv==1.1.1
langchain_aws==0.2.30
langchain-deepseek==0.1.4
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

-e ../common_modules